Add plain non-rails models
