+++ app/repositories/eve/alliance_repository.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Eve
+  class AllianceRepository
+
+  end
+end
